bug fixes, script can now bring up an EC2 instance

+++ New-Instance.ps1
@@ -1,4 +1,5 @@
 ﻿#Requires -Version 5
+#Requires -RunAsAdministrator
 
 using namespace System
 
@@ -15,7 +16,7 @@ param (
 )
 BEGIN
 {
-    $Script:ErrorActionPreference = [System.Management.Automation.ActionPreference]::Stop
+    $Script:ErrorActionPreference = [Management.Automation.ActionPreference]::Stop
     Set-StrictMode -Version 1
 
     if ([string]::IsNullOrWhiteSpace($UserName))
@@ -34,14 +35,14 @@ BEGIN
     {
         throw [ArgumentException]::new("The target AWS region must be specified.", 'Region')
     }
-
+    
     if (-not (Get-Module -Listavailable -Name AWSPowerShell))
     {
         Install-Module -Name AWSPowerShell -Force
     }
     Import-Module AWSPowerShell
 
-    Initialize-AWSDefaults -AccessKey $AccessKey -SecretKey $SecretAccessKey -Region $Region
+    Initialize-AWSDefaultConfiguration -AccessKey $AccessKey -SecretKey $SecretAccessKey -Region $Region
 
     Set-Item WSMan:\localhost\Client\TrustedHosts "*" -Force
 
@@ -49,7 +50,7 @@ BEGIN
     [string] $securityGroup = 'mySecGroup'
     [string] $keyName = 'myKey'
 
-    functon Get-SecurityGroup
+    function New-SecurityGroup
     {
         [CmdletBinding(PositionalBinding = $false)]
         param (
@@ -62,24 +63,29 @@ BEGIN
             throw [ArgumentException]::new("The name of the security group must be specified.", 'SecurityGroupName')
         }
 
-        [string] $myIpRange = '0.0.0.0/0'
-        $groupId = New-EC2SecurityGroup $SecurityGroupName  -Description "Test Security Group"
-        Get-EC2SecurityGroup -GroupNames $SecurityGroupName
+        if (Get-EC2SecurityGroup -GroupName $SecurityGroupName | ? { $_.GroupName -eq $SecurityGroupName })
+        {
+        
+            Write-Host "Security group ""$SecurityGroupName"" exists, skipping."
+            return
+        }
 
-        Write-Verbose 'Opening firewall for ping'
+        [string] $myIpRange = '0.0.0.0/0'
+
+        New-EC2SecurityGroup -GroupName $SecurityGroupName  -Description "Test Security Group"
+        
+        Write-Host 'Opening firewall for ping'
         Grant-EC2SecurityGroupIngress -GroupName $SecurityGroupName -IpPermissions @{IpProtocol = "icmp"; FromPort = -1; ToPort = -1; IpRanges = @($myIpRange)}
         
-        Write-Verbose 'Opening firewall for RDP'
+        Write-Host 'Opening firewall for RDP'
         Grant-EC2SecurityGroupIngress -GroupName $SecurityGroupName -IpPermissions @{IpProtocol = "tcp"; FromPort = 3389; ToPort = 3389; IpRanges = @($myIpRange)}
         Grant-EC2SecurityGroupIngress -GroupName $SecurityGroupName -IpPermissions @{IpProtocol = "udp"; FromPort = 3389; ToPort = 3389; IpRanges = @($myIpRange)}
         
-        Write-Verbose 'Opening firewall for WinRM'
+        Write-Host 'Opening firewall for WinRM'
         Grant-EC2SecurityGroupIngress -GroupName $SecurityGroupName -IpPermissions @{IpProtocol = "tcp"; FromPort = 5985; ToPort = 5986; IpRanges = @($myIpRange)}
-    
-        return $groupId
     }
 
-    function Get-KeyPair
+    function New-KeyPair
     {
         [CmdletBinding(PositionalBinding = $false)]
         param (
@@ -91,56 +97,69 @@ BEGIN
         {
             throw [ArgumentException]::new("The name of key for encryption of Administrator password must be specified.", 'KeyName')
         }
-
-        $keyPair = New-EC2KeyPair -KeyName $KeyName
-        $keyFilePath = [IO.Path]::Combine($PSScriptRoot, "$KeyName.pem")
+        
+        if (Get-EC2KeyPair | ? { $_.KeyName -eq $KeyName })
+        {
+            Write-Host "Key ""$KeyName"" exists, skipping."
+            return
+        }
+        
+        [Amazon.EC2.Model.KeyPair] $keyPair = New-EC2KeyPair -KeyName $KeyName
+        [string] $keyFilePath = [IO.Path]::Combine($PSScriptRoot, "$KeyName.pem")
+        Write-Host "Backing up key: ""$keyFilePath""."
 
         if ( Test-Path $keyFilePath -PathType Leaf )
         {
+            Write-Host "Removing stale file: ""$keyFilePath""."
             Remove-Item -LiteralPath $keyFilePath -Force
         }
-
+        
         "$($keyPair.KeyMaterial)" | Out-File -Encoding ascii -FilePath $keyFilePath
         "KeyName: $($keyPair.KeyName)" | Out-File -Encoding ascii -FilePath $keyFilePath -Append
         "KeyFingerprint: $($keyPair.KeyFingerprint)" | Out-File -Encoding ascii -FilePath $keyFilePath -Append
+    }
 
-        return $keyPair
+    function Get-ImageId
+    {
+        [Object[]]$amiObj = Get-EC2Image -Filter @{ Name="name"; Values="ubuntu*" } -Owner amazon
+        if ($amiObj.Count -lt 1)
+        {
+            throw [Exception]::new('Failed to find ubuntu AMI')    
+        }
+        
+        return $amiObj[0].ImageId 
     }
 }
 PROCESS
 {
-    Get-SecurityGroup -SecurityGroupName $securityGroup
-    $keyPair = Get-KeyPair -KeyName $keyName
+    New-SecurityGroup -SecurityGroupName $securityGroup
+    New-KeyPair -KeyName $keyName
+    [ValidateNotNullOrEmpty()][string] $imageId = Get-ImageId
     
-    [Object[]]$amiObj = Get-EC2Image -Filter @{ Name="name"; Values="ubuntu*" } -Owner amazon
+    [Amazon.EC2.Model.Reservation] $ec2Instance = New-EC2Instance -ImageId $imageId -MinCount 1 -MaxCount 1 -KeyName $keyName -SecurityGroups $securityGroup -InstanceType $instanceType
 
-    if ($amiObj.Count -lt 1)
-    {
-        throw [Exception]::new('Failed to find ubuntu AMI')    
-    }
+    [ValidateNotNullOrEmpty()][string] $instanceId = $ec2Instance.Instances[0].InstanceId
 
-    $ec2Instance = New-EC2Instance -ImageId $amiObj[0].ImageId -MinCount 1 -MaxCount 1 -KeyName $keyPair -SecurityGroups $securityGroup -InstanceType $instanceType
-
-    $instanceId = $ec2Instance.Instances[0].InstanceId
-    $instaceFilter = @{Name = "instance-id"; Values = $instanceId}
     $timer = [Diagnostics.Stopwatch]::StartNew()
-    while ((Get-EC2Instance -Filter $instaceFilter).Instances[0].State.Name -ine 'Running')
+    while ((Get-EC2Instance -InstanceId $instanceId).Instances[0].State.Name -ine 'Running')
     {
-        Write-Verbose -Message "Waiting for instance ""$instanceId"" enter running state."
-        if ($timer.Elapsed -ge [timespan]::FromMinutes(15))
+        Write-Host "Waiting for instance ""$instanceId"" enter running state."
+        if ($timer.Elapsed -ge [timespan]::FromMinutes(30))
         {
             throw [Exception]::new("Timeout exceeded. Instance ""$instanceId"" failed to enter running state.")
         }
         Start-Sleep -Seconds 30
     }
+    Write-Host "Instance ""$instanceId"" is running."
     $timer.Stop()
 
-    $publicDnsName = (Get-EC2Instance -Filter $instaceFilter).Instances[0].PublicDnsName
+    [ValidateNotNullOrEmpty()][string] $publicDnsName = (Get-EC2Instance -InstanceId $instanceId).Instances[0].PublicDnsName
+    Write-Host "Waiting for instance ""$instanceId"" has DNS name:""$publicDnsName""."
 
     $timer = [Diagnostics.Stopwatch]::StartNew()
     while (-not (Test-Connection -Quiet -ComputerName $publicDnsName -Count 1))
     {
-        Write-Verbose -Message "Waiting for ping response from ""$publicDnsName""."
+        Write-Host "Waiting for ping response from ""$publicDnsName""."
         if ($timer.Elapsed -ge [timespan]::FromMinutes(15))
         {
             throw [Exception]::new("Timeout exceeded. ""$publicDnsName"" failed to respond to ping command.")
@@ -148,4 +167,6 @@ PROCESS
         Start-Sleep -Seconds 10
     }
     $timer.Stop()
+    Write-Host "Instance online: ""$publicDnsName""."
 }
+
